+++ src/common/ImageReference.cpp
@@ -46,16 +46,20 @@ std::string ImageReference::string() const {
     return output.str();
 }
 
-std::string ImageReference::normalize() const {
-    auto output = std::stringstream{};
-    output << getFullName();
-    if (!digest.empty()){
-        output << "@" << digest;
+/**
+ * Normalizing a reference means clearing the tag if the digest is also present.
+ * This is useful to reproduce Docker's behavior, which completely ignores the tag
+ * when a digest is given. Podman and Buildah also implement this behavior for
+ * compatibility.
+ * See for reference: https://github.com/containers/common/pull/579,
+ * https://github.com/containers/common/blob/v0.47.4/libimage/normalize.go
+ */
+ImageReference ImageReference::normalize() const {
+    auto output = *this;
+    if (!digest.empty() && !tag.empty()){
+        output.tag.clear();
     }
-    else if (!tag.empty()){
-        output << ":" << tag;
-    }
-    return output.str();
+    return output;
 }
 
 bool operator==(const ImageReference& lhs, const ImageReference& rhs) {
@@ -71,6 +75,22 @@ std::ostream& operator<<(std::ostream& os, const ImageReference& imageReference)
     return os;
 }
 
+/**
+ * Creates a string which can be used to construct a filesystem path to the location
+ * of squashfs and metadata files for an image within a Sarus local repository.
+ * The tag has higher priority than the digest because:
+ *
+ * - files for image pulled by tag are separate than files for images pulled by digest
+ * - ImageManager::pullImage() completes the reference with a digest in case a digest is
+ *   not supplied vie the CLI. This is consistent with the behavior of Docker.
+ *
+ * Therefore, if getUniqueKey() gave higher priority to the digest, when a tag+digest
+ * reference is used in the CLI, the consumers of the uniqueKey would fail to find the
+ * files named after the tag, or worse delete the files named after the digest.
+ *
+ * This situation could be streamlined in the future by only storing images by digest and
+ * having ImageStore resolve tags to the correct images/files stored.
+ */
 std::string ImageReference::getUniqueKey() const {
     auto output = std::stringstream{};
     output << getFullName();